회의예약edit 수정

+++ src/controllers/meetingController.ts
@@ -21,7 +21,7 @@ class meetingController {
 
         // const member = await dbs.Participants.findUser()
         const result = await Promise.all(_.map(list, async (meeting: any) => {
-            const membersObj = await dbs.Participant.findAllUser(meeting.id)
+            const membersObj =  await dbs.Participant.findAllUser(meeting.id)
 
             const memberNameList = await Promise.all(_.map(membersObj, async (member: any) => {
                 const memberInfo = await userController.getUserInfo(member.user_id)
@@ -35,7 +35,7 @@ class meetingController {
                     "type": "section",
                     "text": {
                         "type": "mrkdwn",
-                        "text": `📢*${meeting.title}* \n\n \`\`\`${moment(meeting.date, 'yyyy-MM-DD').format('yyyy-MM-DD')} ${moment(meeting.start, 'HH:mm:ss').format("HH:mm")} — ${moment(meeting.end, 'HH:mm:ss').format("HH:mm")}\`\`\` 참석자 : ${memberNameList}\n\n`
+                        "text": `📢*${meeting.title}* \n\n 참석자 : ${_.map(memberNameList, (name:any)=>{return ' ' + name })}\n\n \`\`\`${moment(meeting.date, 'yyyy-MM-DD').format('yyyy-MM-DD')} ${moment(meeting.start, 'HH:mm:ss').format("HH:mm")} — ${moment(meeting.end, 'HH:mm:ss').format("HH:mm")}\`\`\` `
                     },
 
                     "accessory": {
@@ -76,11 +76,9 @@ class meetingController {
     }
 
     createMeetingForm(data: any, user: any) {
-        // const values = data.state.values;
-        // const blocks = data.blocks;
-        const createMeeting = {
+               const createMeeting = {
             user_id: user.id,
-            room_number: data.roomNumber,
+            room_number: data.room_number,
             title: data.title,
             description: data.description,
             date: data.date,
@@ -88,16 +86,6 @@ class meetingController {
             end: data.end,
             members: data.members
         }
-        // const createMeeting = {
-        //     user_id: user.id,
-        //     room_number: values[blocks[0].block_id].room_number.selected_option.value,
-        //     title: values[blocks[1].block_id].title ? values[blocks[1].block_id].title.value : '',
-        //     description: values[blocks[2].block_id].description.value !== null ? values[blocks[2].block_id].description.value : '',
-        //     date: values[blocks[3].block_id].selected_date.selected_date,
-        //     duration: values[blocks[4].block_id].meeting_duration.selected_option.value,
-        //     end: values[blocks[4].block_id].meeting_end.selected_option ? values[blocks[4].block_id].meeting_end.selected_option.value : '',
-        //     members: values[blocks[5].block_id].participant_list.selected_users,
-        // }
 
         return createMeeting
     }
@@ -138,14 +126,9 @@ class meetingController {
 
     }
 
-    // sendResponse= async (booking_id: string, user: any, trigger_id: string) => {
-    //     const result = axios.post()
-    // }
-
 
     deleteMeeting = async (meeting_id: string, user: any, trigger_id: string) => {
         const deleteMeeting = await dbs.Meeting.destroy({id: meeting_id})
-
 
         return deleteMeeting
 
@@ -158,48 +141,51 @@ class meetingController {
 
     }
 
-    editMeeting = async (view: any, meeting_id: any, user: any) => {
+    editMeeting = async (data: any, meeting_id: number, user: any) => {
 
         return dbs.Meeting.getTransaction(async (transaction: Transaction) => {
-            const values = view.state.values;
-            const blocks = view.blocks;
+
+            const meetingForm = this.createMeetingForm(data, user);
+
+            const hasMeeting = await dbs.Meeting.hasMeetingAtTime(new Date(meetingForm.date), meetingForm.room_number, meetingForm.start, meetingForm.end);
 
 
-            const meetingInfo = {
-                room_number: values[blocks[1].block_id].room_number.selected_option.value,
-                title: values[blocks[2].block_id].title.value,
-                description: values[blocks[3].block_id].description.value,
-                date: values[blocks[4].block_id].selected_date.selected_date,
-                // start: values[blocks[5].block_id].start.selected_option.value,
-                // end: values[blocks[5].block_id].end.selected_option.value,
-                // state: eBookingState.Modified,
-            }
+            _.forEach(hasMeeting, (meeting:any, idx:number)=>{
+              if(meeting.id == meeting_id){
+                   hasMeeting.splice(idx,1)
+              }
+            })
 
-            const participantList: any = [];
-            const participantArr = values[blocks[6].block_id].participant_list.selected_users;
+            if (hasMeeting && hasMeeting.length === 0) {
+
+                const participantList: any = [];
+                const participantArr = data.members;
 
 
-            const meeting = await dbs.Meeting.update(meetingInfo, {id: meeting_id}, transaction);
+                await dbs.Meeting.update(meetingForm, {id: meeting_id}, transaction);
 
 
-            for (let i = 0; i < participantArr.length; i++) {
+                for (let i = 0; i < participantArr.length; i++) {
 
-                let obj = {
-                    user_id: participantArr[i],
-                    meeting_id: meeting_id
-                    // name:view.username
+                    let obj = {
+                        user_id: participantArr[i],
+                        meeting_id: meeting_id
+                        // name:view.username
+                    }
+                    participantList.push(obj)
                 }
-                participantList.push(obj)
+
+
+                // throw new Error()
+                await dbs.Participant.destroy({meeting_id: meeting_id});
+
+                await dbs.Participant.bulkCreate(participantList, transaction);
+            }else {
+
+                throw new Error('이미 등록된 예약이 있습니다.')
             }
-
-
-            // throw new Error()
-            console.log(participantList)
-            await dbs.Participant.destroy({meeting_id: meeting_id});
-
-            const result = await dbs.Participant.bulkCreate(participantList, transaction);
-
         })
+
 
 
     }
